Avoid refitting the mosaic on resize when it is hidden
Refitting hidden mosaic is pointless and it does not work anyway (the
images end up with 0×0 dimensions).  When the gallery is revealed and the
viewport size has changed while it was hidden, your code must invoke a
'jqMosaicRefit' event on the main div (DIV.mosaic), otherwise it will no
longer be correctly fit to it.  This could possibly be made automatic
with MutationObserver, but for the time being, this should suffice.